+++ ddi/tbor/types/tests/commands/part_init/mod.rs
@@ -26,11 +26,7 @@
 
 #![cfg(feature = "emu")]
 
-use azihsm_ddi_tbor_types::PolicyKeyKind;
 use azihsm_ddi_tbor_types::SessionType;
-use azihsm_ddi_tbor_types::MACH_SEED_LEN;
-use azihsm_ddi_tbor_types::PART_POLICY_LEN;
-use azihsm_ddi_tbor_types::POTA_THUMBPRINT_LEN;
 use azihsm_ddi_tbor_types::PSK_LEN;
 use azihsm_ddi_tbor_types::SATA_THUMBPRINT_LEN;
 
@@ -45,6 +41,16 @@ mod sd_config;
 
 pub(crate) const CO: u8 = 0;
 
+// Wire-format `PartInit` fixtures live in `crate::harness::part_policy`
+// so the hw-eligible `default_psk_gate` tests can reuse them without
+// ungating this emu-only module. Re-exported at the same paths the
+// submodules already import through (`super::known_good_part_policy`,
+// `super::mach_seed`, etc.).
+pub(crate) use crate::harness::known_good_part_policy;
+pub(crate) use crate::harness::mach_seed;
+pub(crate) use crate::harness::part_policy_with_pota;
+pub(crate) use crate::harness::pota_thumbprint;
+
 /// Non-default 32-byte CO PSK used so PartInit clears the
 /// default-PSK-gate.  Pinned to a fixed value so the smoke test is
 /// fully deterministic.
@@ -52,67 +58,6 @@ pub(crate) const ROTATED_CO_PSK: [u8; PSK_LEN] = [
     0xA1, 0xA2, 0xA3, 0xA4, 0xA5, 0xA6, 0xA7, 0xA8, 0xA9, 0xAA, 0xAB, 0xAC, 0xAD, 0xAE, 0xAF, 0xB0,
     0xB1, 0xB2, 0xB3, 0xB4, 0xB5, 0xB6, 0xB7, 0xB8, 0xB9, 0xBA, 0xBB, 0xBC, 0xBD, 0xBE, 0xBF, 0xC0,
 ];
-
-/// Build a 484-byte unified `PartPolicy` blob that passes
-/// `azihsm_fw_hsm_core::ddi::tbor::policy::from_bytes`.  Layout mirrors
-/// the canonical wire format defined in
-/// `fw/core/ddi/tbor/types/src/policy.rs`: POTA + SATA trust anchors are
-/// populated Ecc384 keys; SAPOTA + backing-partition keys are left
-/// absent (zero `len`); flags are clear; `info` is filled.
-pub(crate) fn known_good_part_policy() -> [u8; PART_POLICY_LEN] {
-    const OFF_POTA: usize = 2;
-    const OFF_SATA: usize = 102;
-    const OFF_FLAGS: usize = 418;
-    const OFF_INFO: usize = 419;
-
-    // Write an Ecc384 (kind 0) raw X‖Y pubkey at `off` (no SEC1 prefix).
-    fn write_pubkey(bytes: &mut [u8], off: usize, fill: u8) {
-        bytes[off..off + 2].copy_from_slice(&PolicyKeyKind::Ecc384.0.to_le_bytes());
-        bytes[off + 2..off + 4].copy_from_slice(&96u16.to_le_bytes());
-        for (i, b) in bytes[off + 4..off + 4 + 96].iter_mut().enumerate() {
-            *b = (fill.wrapping_add(i as u8)) | 0x80;
-        }
-    }
-
-    let mut bytes = [0u8; PART_POLICY_LEN];
-    bytes[0] = 1; // version major
-    bytes[1] = 0; // version minor
-    write_pubkey(&mut bytes, OFF_POTA, 0x10);
-    write_pubkey(&mut bytes, OFF_SATA, 0x20);
-    // SAPOTA + backup-part pubkeys left absent (len 0).
-    bytes[OFF_FLAGS] = 0;
-    for b in bytes[OFF_INFO..OFF_INFO + 64].iter_mut() {
-        *b = 0xAB;
-    }
-    bytes
-}
-
-/// Like [`known_good_part_policy`] but with a caller-supplied **real**
-/// `POTAPubKey` (raw P-384 `X ‖ Y`, 96 bytes), so `PartFinal` can validate
-/// a PTA certificate chain anchored to it.
-pub(crate) fn part_policy_with_pota(pota_raw: &[u8; 96]) -> [u8; PART_POLICY_LEN] {
-    const OFF_POTA: usize = 2;
-    let mut bytes = known_good_part_policy();
-    // POTA slot layout: kind(2) ‖ len(2) ‖ data(96); overwrite the data.
-    bytes[OFF_POTA + 4..OFF_POTA + 4 + 96].copy_from_slice(pota_raw);
-    bytes
-}
-
-pub(crate) fn mach_seed() -> [u8; MACH_SEED_LEN] {
-    let mut v = [0u8; MACH_SEED_LEN];
-    for (i, b) in v.iter_mut().enumerate() {
-        *b = 0x40 + i as u8;
-    }
-    v
-}
-
-pub(crate) fn pota_thumbprint() -> [u8; POTA_THUMBPRINT_LEN] {
-    let mut v = [0u8; POTA_THUMBPRINT_LEN];
-    for (i, b) in v.iter_mut().enumerate() {
-        *b = 0x80 ^ i as u8;
-    }
-    v
-}
 
 pub(super) fn sata_thumbprint() -> [u8; SATA_THUMBPRINT_LEN] {
     let mut v = [0u8; SATA_THUMBPRINT_LEN];
